\name{checkInstalled}
\alias{checkInstalled}
\title{check whether external program is installed}
\description{This function checks whether an external program is installed on the machine.}
\usage{checkInstalled(commandName)}
\arguments{
  \item{commandName}{character string: name of the program/command to check}
}

\value{logical: whether the program is installed.}

\author{Jannis v. Buttlar}





