\name{convertArgs2String}
\alias{convertArgs2String}
\title{Save function argument values to a character string}
\description{convertArgs2Strings saves function argument settings to a character string.
details <<
convertArgs2Strings saves function argument settings to a character string.
This is used in the ncdf routines that save these argument values as
attributes of the target ncdf files.}
\usage{convertArgs2String()}
\value{character string: arguments given to function converted into a string.         }

\author{Jannis v. Buttlar}





