\name{getSysinfo}
\alias{getSysinfo}
\title{extract system information}
\description{convenience function to compile some information about the system used.}
\usage{getSysinfo()}

\value{character string with all version and system information of the current R system}

\author{Jannis v. Buttlar}





