\name{getVecInfo}
\alias{getVecInfo}
\title{Compute vector summary statistics}
\description{This function computes several summary statistics of a vector.}
\usage{getVecInfo(x)}
\arguments{
  \item{x}{
}
}

\value{vector with the vector statistics (min,  max, mean,  sdev, range, ratio na, ratio na inner,  ratio INF)}

\author{Jannis v. Buttlar}





