\name{indexDimVecs2Matrix}
\alias{indexDimVecs2Matrix}
\title{Transform integer indices to an index matrix}
\description{Transform integer indices to an index matrix. The input have to be
index vector for each dimension of the target array.
\code{\link{Extract}},  \code{\link{indexVec2Matrix}}}
\usage{indexDimVecs2Matrix(...)}\arguments{
  \item{\dots}{integer vectors: indices to use for the different dimensions}
}

\value{Index matrix}

\author{Jannis v. Buttlar}





