\name{indexVec2Matrix}
\alias{indexVec2Matrix}
\title{Transform a vector index to an index matrix}
\description{Transform an index vector into a index matrix.
\code{\link{Extract}}}
\usage{indexVec2Matrix(index, dim)}\arguments{
  \item{index}{vector index}
  \item{dim}{length of the two dimensions of the array to index. Identical to
the result of dim(array).}
}

\value{index array}

\author{Jannis v. Buttlar}





