\name{JBTools-package}
\alias{JBTools-package}
\alias{JBTools}
\docType{package}
\title{Collection tools for R}
\description{Collection of tools and functions used by other packages.}
\details{
\tabular{ll}{Package: \tab JBTools\cr
Title: \tab Collection tools for R\cr
Version: \tab 0.7.0.0\cr
Date: \tab 2011-02-15\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab foreach, parallel, gplots, colorspace, plotrix\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 1.10.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


