\name{labelMargins}
\alias{labelMargins}
\title{Label the margins of a plot}
\description{Writes equidistant labels to the margins of a plot.}
\usage{labelMargins(labels, side = 1, ...)}
\arguments{
  \item{labels}{character vector: labels to use}
  \item{side}{integer: side of the plot to label}
  \item{\dots}{further arguments passed to the plotting routines}
}

\value{nothing is returned}

\author{Jannis v. Buttlar}





