\name{MEF}
\alias{MEF}
\title{compute the MEF}
\description{Calculates the modelling efficiency (MEF) of two arrays }
\usage{MEF(prediction, observation)}
\arguments{
  \item{prediction}{numeric: array 1}
  \item{observation}{numeric: array 2}
}

\value{MEF of the two vectors}

\author{Jannis v. Buttlar}





