\name{printStatus}
\alias{printStatus}
\title{print and save a status report}
\description{This function prints a given string as a status report to the screen together
with the current time.}
\usage{printStatus(string.in)}
\arguments{
  \item{string.in}{character string: status message to print}
}



\author{Jannis v. Buttlar}





