\name{rangeZeroEqui}
\alias{rangeZeroEqui}
\title{Compute a zero centered equi-sided range}
\description{rangeZeroEqui computes a zero centered equi-sided range,  i.e. it
the maximum absolute value and returns this as a positive and a
negative value. }
\usage{rangeZeroEqui(x)}
\arguments{
  \item{x}{input vector}
}

\value{vector}

\author{Jannis v. Buttlar}





