\name{RMSE}
\alias{RMSE}
\title{Compute the residual mean square error (RMSE)}
\description{Calculates the residual mean square error (RMSE) of two arrays}
\usage{RMSE(prediction, observation)}\arguments{
  \item{prediction}{numeric: array 1}
  \item{observation}{numeric: array 2}
}

\value{RMSE of the two vectors}

\author{Jannis v. Buttlar}





