\name{seqLongest}
\alias{seqLongest}
\title{find longest sequence of TRUEs}
\description{The function determines the first position and the length of the longest
uninterupted series of TRUEs in a logical sequence. This for example can be
used to find gapless positions in a series}
\usage{seqLongest(array_in, na.rm = FALSE)}
\arguments{
  \item{array_in}{logical vector: input vector}
  \item{na.rm}{logical: if TRUE, NAs are tretaed as FALSE}
}



\author{Jannis v. Buttlar}





