\name{getSysinfo}
\alias{getSysinfo}
\title{Compile system information}
\description{getSysinfo is a convenience function to compile some information about the
computing system and environment used.}
\usage{getSysinfo()}\details{The function is mainly used to save the system environment information
in ncdf files containing the results of some calculations. }
\value{character string with all version and system information of the current R system}

\author{Jannis v. Buttlar}





