\name{JBTools-package}
\alias{JBTools-package}
\alias{JBTools}
\docType{package}
\title{Collection of miscellaneous tools and helpers for R.}
\description{Collection of tools and functions used by other packages of the package author.
Some functions may be of use to others for general tasks, others are mainly used by the
spectral.methods and ncdf.tools package. All code is under constant development and errors
or bugs sometimes occour when the functions are used in new contexts and scenarios which
have not yet been tested. Please report all such errors to the mailing lists on R-Forge and
the author will try to fix them, time permitting.}
\details{
\tabular{ll}{Package: \tab JBTools\cr
Title: \tab Collection of miscellaneous tools and helpers for R.\cr
Version: \tab 0.7.0.1\cr
Date: \tab 2011-02-15\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab foreach, parallel, gplots, colorspace, plotrix, doMC\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.14.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


