\name{JBTools-package}
\alias{JBTools-package}
\alias{JBTools}
\docType{package}
\title{JBs tools and helper functions.}
\description{Collection of tools and functions used by other packages of the package author. Most functions are specific for their usage in these packages but some are of general use for other applications.}
\details{
\tabular{ll}{Package: \tab JBTools\cr
Title: \tab JBs tools and helper functions.\cr
Version: \tab 0.7.2.4\cr
Date: \tab 2014-01-01\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab foreach, parallel, gplots, colorspace, plotrix, doMC\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.14.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


