\name{JBTools-package}
\alias{JBTools-package}
\alias{JBTools}
\docType{package}
\title{JBs tools and helper functions.}
\description{Collection of several tools and helper functions used across the other packages of J. Buttlar (ncdf.tools and spectral.methods).}
\details{
\tabular{ll}{Package: \tab JBTools\cr
Title: \tab JBs tools and helper functions.\cr
Version: \tab 0.7.2.5\cr
Date: \tab 2014-01-01\cr
Author: \tab Jannis v. Buttlar\cr
Maintainer: \tab Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>\cr
Imports: \tab foreach, parallel, gplots, colorspace, plotrix\cr
Suggests: \tab doMC\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.14.0)\cr}
}
\author{Jannis v. Buttlar}

\keyword{ package }


