\name{printStatus}
\alias{printStatus}
\title{Print a status message}
\description{This function prints a given string as a status report to the screen together
with the current time.}
\usage{printStatus(string.in)}\arguments{
  \item{string.in}{character string: status message to print}
}

\value{Nothing is returned but a message is printed to the screen.}

\author{Jannis v. Buttlar}





