\name{whichClosest}
\alias{whichClosest}
\title{Find closest matches in two vectors}
\description{Function to find closest matches of vector A in vector B and return
the respective indices.}
\usage{whichClosest(x, x.match, arr.ind = FALSE)}\arguments{
  \item{x}{numeric vector: values that should be found in x.match}
  \item{x.match}{numeric vector: values the values of x should be matched against}
  \item{arr.ind}{
}
}

\value{integer vector: indices of the closest element of x.match to each entry in x}

\author{Jannis v. Buttlar}





